Fixed an issue when reading from queue file.
Issues including:
> Error reading queue invalid cast exception
> Format part of download entry is not a number

These issues are present because a change was made to include sites other than YouTube.com. YouTube URLs when put through youtube-dl are always numeric in nature. Other sites (twitter, facebook) don't have numeric file formats.

To put it simply the issue occurred when a section of code was excepting an integer and got a string.